requirements.txt
argh==0.26.2
euclid3==0.1
pathtools==0.1.2
prettytable==0.7.2
pypng==0.0.19
PyYAML==5.3.1
regex==2019.12.20
six==1.14.0
solidpython==0.4.8
watchdog==0.10.2
